+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.3]
+
+### Fixed
+
+- `bydi-mock--volatile` no longer returns the resut of `memq` which
+  can mess up tests. Instead it returns `t` or `nil`.g
+
 ## [0.6.2]
 
 ### Added